pyproject.toml
[project]
name = "ai-tutor-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.8.2",
  "sqlmodel>=0.0.21",
  "sqlalchemy>=2.0.35",
  "psycopg[binary]>=3.2.1",
  "httpx>=0.27.0"
]